% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFloralNesting.R
\name{computeFloralNesting}
\alias{computeFloralNesting}
\title{Compute floral and nesting resource values for flower-visiting insects}
\usage{
computeFloralNesting(landuseMap, edgesMap, unitEdges = "m", widthEdges,
  landuseCodes, bees, num_floral, florNestInfo, codeEdges, cell.size = 25)
}
\arguments{
\item{landuseMap}{the map of the landscape containing the landuse categories, given as a raster}

\item{edgesMap}{the map containing the length of field edges in each cell, given as a raster or a rasterstack in the case of multiple edge types}

\item{landuseCodes}{the matching between landuse codes and landuse category names}

\item{bees}{the vector of bee species}

\item{num_floral}{the number of floral periods}

\item{florNestInfo}{a list containing the floral and nesting information, with 4 elements: 1) the floral coverage, 2) bumblee bee info, i.e. floral value and attractiveness, 3) honeybee info and 4) solitary bee info}

\item{codeEdges}{a vector containing the landuse codes for the edge types (e.g. grassy field edge or sown wildflower strip)}

\item{cell.size}{a number in meters corresponding to the widht and breadth of the input raster cell}
}
\value{
a list containing two arrays: one of dimension nb of bees containing the nesting qualities, and one of dimension nb of bees * nb of periods containing the floral values for each species and each period
}
\description{
Function that computes the floral and nesting arrays according to the bees that are accounted for
}
