% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgearea_example.R
\docType{data}
\name{edgearea_example}
\alias{edgearea_example}
\title{Edge area data}
\format{An object of class \code{RasterLayer} of dimension 160 x 160 x 1.}
\usage{
edgearea_example
}
\description{
A 4x4 km raster layer from southern Sweden with 2014 permanent field edge area information (in sqm).
Created by Peter Olsson, CEC, Lund University, from information obtained from the Swedish Board of Agriculture,
for use in ecosystem service models.
}
\keyword{datasets}
