% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.func.R
\name{growth.func}
\alias{growth.func}
\title{Defines pollinator population growth function}
\usage{
growth.func(R, a, b, max)
}
\arguments{
\item{R}{: the resources as a matrix}

\item{a}{: the inflexion point in the logistic curve (i.e the median)}

\item{b}{: the slope of the growth}

\item{max}{: the asymptote}

\item{Nindiv}{: the current population size as a matrix}
}
\description{
Generic growth function for bees population according to the available resources, using the inverse logit function
}
\references{
Haeussler J, Sahlin U, Baey C, Smith HG, Clough Y (2017) Predicting pollinator population size and pollination ecosystem service responses to enhancing floral and nesting resources. Ecology and Evolution, 7: 1898-1908.\url{http://dx.doi.org/10.1002/ece3.2765}
}
