% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerncalc.R
\name{kerncalc}
\alias{kerncalc}
\title{Exponential kernel computation}
\usage{
kerncalc(beta, cell.size, maxSize = Inf, decaycut = 0.99)
}
\arguments{
\item{beta}{the mean of the exponential distribution}

\item{cell.size}{the size of a lattice's cell in meters}

\item{maxSize}{The maximum alowable size of the resulting matrix.}

\item{decaycut}{the quantile used to cut the tails of the kernel}
}
\description{
Exponential kernel computation to be used as an input of the latfordisp function.
}
