% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerncalc_mort.R
\name{kerncalc_mort}
\alias{kerncalc_mort}
\title{Exponential kernel computation, with mortality}
\usage{
kerncalc_mort(beta, cell.size, maxSize = Inf, decaycut = 0.99,
  surv.per.m = 0.9)
}
\arguments{
\item{beta}{the mean of the exponential distribution}

\item{cell.size}{the size of a lattice's cell in meters}

\item{maxSize}{The maximum alowable size of the resulting matrix.}

\item{decaycut}{the quantile used to cut the tails of the kernel}

\item{surv.per.m}{proportion of individuals surviving per meter travelled}
}
\description{
Exponential kernel computation to be used as an input of the latfordisp function.
}
