% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerncalc_old.R
\name{kerncalc_old}
\alias{kerncalc_old}
\title{Exponential kernel computation, old version}
\usage{
kerncalc_old(beta, cell.size, maxSize = Inf, decaycut = 0.99)
}
\arguments{
\item{beta}{the mean of the exponential distribution}

\item{cell.size}{the size of a lattice's cell in meters}

\item{maxSize}{The maximum alowable size of the resulting matrix.}

\item{decaycut}{the quantile used to cut the tails of the kernel}
}
\description{
An older version of kerncalc, should not be used. Kept for documentation.
}
