% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse_example.R
\docType{data}
\name{landuse_example}
\alias{landuse_example}
\title{Land-use data}
\format{An object of class \code{RasterLayer} of dimension 160 x 160 x 1.}
\usage{
landuse_example
}
\description{
A 4x4 km raster layer from southern Sweden with 2014 land-use crop-cover information.
Created by Peter Olsson, CEC, Lund University, from information obtained from the Swedish Board of Agriculture,
for use in ecosystem service models.
}
\keyword{datasets}
