% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latfordisp.R
\name{latfordisp}
\alias{latfordisp}
\title{Apply kernels to model foraging or dispersal in space}
\usage{
latfordisp(N, alpha, decay)
}
\arguments{
\item{N}{the matrix of foragers (source units)}

\item{alpha}{the matrix of resources}

\item{decay}{the dispersal kernel, for example produced by kerncalc}
}
\value{
a list containing :
\describe{
  \item{Racc_per_forager}{the number of queens at the end of each period and at the beginning of the next season}
  \item{visitation_rate}{the visitation rates per cell}
  \item{distance_weighted_resources}{matrix of resources weighted by the kernal}
 }
}
\description{
For foragers occupying a or more cells of a matrix, applies a spatial kernel
weighted by the available resources such that foragers appear in cells that are
both close-by and have higher resources.
}
