% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noz.R
\name{noz}
\alias{noz}
\title{No zero helper function}
\usage{
noz(x)
}
\arguments{
\item{x}{a matrix or a vector}
}
\value{
same as x, that is a matrix or a vector
}
\description{
Replace negative values by zeros and rescale such that the same sum is preserved
}
