% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/springbarley_biocontrol.R
\name{springbarley_biocontrol}
\alias{springbarley_biocontrol}
\title{Springbarley aphid biocontrol model function}
\usage{
springbarley_biocontrol(useraster = TRUE, cSMD, mask, parameters,
  reclassin_PGL, reclassin_PAC, N00 = "usedefault")
}
\arguments{
\item{useraster}{whether to use a raster, if FALSE the function assumes a matrix.}

\item{cSMD}{land-use data, provided as a raster or a matrix.}

\item{mask}{same format as cSMD but containing binary values, 1 = produce a prediction, 0 = ignore}

\item{parameters}{list of two dataframes containing parameter values}

\item{reclassin_PGL}{matrix indicating for a sequence of land-use codes the grasslands (1) and the non-grassland uses (0)}

\item{reclassin_PAC}{matrix indicating for a sequence of land-use codes the annual land-uses (1) and the permanent ones (0)}

\item{N00}{the density of colonizing aphids, does not vary in space, if set to usedefault its value is taken from OTHER_PAR.}
}
\value{
a list containing :
\describe{
  \item{mc1}{mortality phase 1}
  \item{mc2}{mortality phase 2}
  \item{ad}{aphid days, calculated as N00*(1-mc1)*(exp((f-mc2)*T)-1)/(f-mc2)}
  \item{cd}{crop loss in kg.ha-1, calculated as 12.7*ad^.66}
 }
}
\description{
Uses spatial data on floral and nesting quality to compute flower visitation and population dynamics for wild pollinators for a given year.
The function can be used for mutliple years by running a loop and using as input the output of the previous model (see example).
useraster=TRUE,cSMD,mask,parameters,landscvals,reclassin_PGL,reclassin_PAC,N00="usedefault"
}
\examples{
PRED_PAR <- read.table("C:/a1workmain/ES_synthesis/BC_springbarley/yann/PRED_PAR.txt")
OTHER_PAR <- read.table("C:/a1workmain/ES_synthesis/BC_springbarley/yann/OTHER_PAR.txt")
parameters.bcsb<-list("PRED_PAR"=PRED_PAR,"OTHER_PAR"=OTHER_PAR)
reclassin_PAC<-as.matrix(reclassin_info[,c(1,2,4)])
reclassin_PGL<-as.matrix(reclassin_info[,c(1,2,5)])
biocontrol_pred<-springbarley_biocontrol(cSMD=cSMD_Sk2020,mask=landsmask,
parameters=parameters.bcsb,reclassin_PGL=reclassin_PGL,reclassin_PAC=reclassin_PAC)
}
\references{
Jonsson, M., Bommarco, R., Ekbom, B., Smith, H. G., Bengtsson, J., Caballero-Lopez, B., Winqvist, C.
and Olsson, O. (2014), Ecological production functions for biological control services in agricultural landscapes.
Methods Ecol Evol, 5: 243-252. \url{http://dx.doi.org/10.1111/2041-210X.12149}
}
